Add toolbar icons for replace control
This commit adds two toolbar icons for replacing all filter/search occurences
with the replacement text either in the selection or in the document.

This commit also adds a toolbar icon to clear the replacement text in
the text control like in the search text control.

This commit considers the cases for case sensitive search and for case
insensitive search.

This commit also remaps the replacement key binding from Ctrl+h to
Ctrl+k because Ctrl+h is superseded by backspace on Linux.

This commit also moves the lowering of the search string directly into
the sys->searchstring so that lowering must be only done once and not in
every call of search and replace functions.