Add toolbar icons for replace control
This commit adds two toolbar icons for replacing all filter/search occurences
with the replacement text either in the selection or in the document.

This commit also adds a toolbar icon to clear the replacement text in
the text control like in the search text control.

This commit considers the cases for case sensitive search and for case
insensitive search and adds overloaded functions for the two cases.
The functions are restructured so that strings are lowered only once
where needed.

This commit also remaps the replacement key binding from Ctrl+h to
Ctrl+k because Ctrl+h is superseded by backspace on Linux.