Copy HTML to clipboard
Create a composite data object that is placed to the clipboard.

The composite data object consists of
- a data object for HTML, either
  - a custom data object in wxGTK containing the HTML string
    or
  - a HTML data object in other toolkits (MSW, OSX)
  and
- a plain text data object (string).

Please note that the HTML data is within a custom data object for wxGTK as
a HTML data object keeps causing a buffer overflow (wxWidgets implementation)
when one tries to paste from the clipboard.

The HTML data object is marked as prefered.