Copy HTML to clipboard
Create a composite data object that is placed to the clipboard.

The composite data object consists of
- a custom data object (prefered) in wxGTK containing the HTML string
- a HTML data object in other toolkits (MSW, OSX)
- a plain text data object (string).

Please note that the HTML data is within a custom data object for wxGTK as
a HTML data object keeps causing a buffer overflow (wxWidgets implementation)
when one tries to paste from the clipboard.