Consider TextEdit mode when copying
If the selected cell is in TextEdit mode, then only the selected part of
the cell text shall be pasted if paste operation is done after the copy
operation.

This can be reached by making the text string in the clipboard different
from the `sys->clipboardcopy` string so that TreeSheets will not take the
complete cell from the internal cell clipboard.