Capture all char/key events for all keys in text controls
Normally the navigations keys are automatically assigned for dialog navigation
in wxWidgets. In TreeSheets we want to be able to navigate with these keys in
the text controls, so add this style.

Also preserve specific behaviour when Escape key is pressed.

See <https://docs.wxwidgets.org/latest/classwx_window.html> for more
details.