Provide similiar experience when Return key is pressed
This provides a similiar experience when Return key is pressed
in the search and the replace text control.

This implements it for MSW and the other ports.

Also there is now a keybinding to jump directly into the replace text
control.

Also there is now an additional check so that Ctrl+letter char events
will not be translated into ASCII control characters (see wxKeyEvent
documentation).