Place Drag and Drop data structures per document
When a page gets closed, the destructor of DropTarget is called
which also frees the associated data. If we maintain only one DropTarget
for all pages, it tries to free the data multiple times, resulting in a
crash.

To avoid this, keep a Drag and Drop data structure in each document.